requirements.txt
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.6.0
